+++ src/demo/SdkMount.tsx
@@ -17,6 +17,10 @@ import * as Onfido from '../index'
 */
 
 const Onfido = window.Onfido
+/* 
+  NOTE: For Auth local build, if you get a TypeError in the console saying "Cannot read property 'init' of undefined" 
+        try using a different import method like import * as Onfido from '../index'
+*/
 
 type Props = {
   options: SdkOptions | UIConfigs