+++ src/demo/demoUtils.ts
@@ -24,6 +24,7 @@ export type QueryParams = {
   forceCrossDevice?: StringifiedBoolean
   hideOnfidoLogo?: StringifiedBoolean
   language?: 'customTranslations' | SupportedLanguages
+  customWelcomeScreenCopy?: StringifiedBoolean
   link_id?: string
   docVideo?: StringifiedBoolean
   faceVideo?: StringifiedBoolean
@@ -152,7 +153,23 @@ export const getInitSdkOptions = (): SdkOptions => {
       ? SAMPLE_LOCALE
       : queryParamToValueString.language
 
-  const steps: Array<StepConfig> = [{ type: 'welcome' }]
+  const steps: Array<StepConfig> = []
+
+  if (queryParamToValueString.customWelcomeScreenCopy === 'true') {
+    steps.push({
+      type: 'welcome',
+      options: {
+        title: 'Open your new bank account',
+        descriptions: [
+          'To open a bank account, we will need to verify your identity.',
+          'It will only take a couple of minutes.',
+        ],
+        nextButton: 'Verify Identity',
+      },
+    })
+  } else {
+    steps.push({ type: 'welcome' })
+  }
 
   if (queryParamToValueString.showAuth === 'true') {
     steps.push({ type: 'auth', options: { retries: 10 } })