+++ test/pageobjects/Welcome.js
@@ -1,9 +1,6 @@
 import BasePage from './BasePage.js'
 import { verifyElementCopy } from '../utils/mochaw'
-import {
-  testFocusManagement,
-  elementCanReceiveFocus,
-} from '../utils/accessibility'
+import { testFocusManagement } from '../utils/accessibility'
 import { Key } from 'selenium-webdriver'
 
 class Welcome extends BasePage {
@@ -36,18 +33,22 @@ class Welcome extends BasePage {
     verifyElementCopy(this.title(), copy.welcome.title)
   }
 
+  async verifyCustomTitle() {
+    verifyElementCopy(this.title(), 'Open your new bank account')
+  }
+
   async verifyFocusManagement() {
     testFocusManagement(this.title(), this.driver)
   }
 
   async verifySubtitle(copy) {
-    verifyElementCopy(this.subtitle(), copy.welcome.doc_video_subtitle)
+    verifyElementCopy(this.subtitle(), copy.welcome.subtitle)
   }
 
-  async verifyInstructions(copy) {
+  async verifyDescriptions(copy) {
     verifyElementCopy(
-      this.instructions(),
-      `${copy.welcome.list_header_doc_video}\n${copy.welcome.list_item_doc}\n${copy.welcome.list_item_selfie}`
+      this.text(),
+      `${copy.welcome.list_header_webcam}\n${copy.welcome.list_item_doc}\n${copy.welcome.list_item_selfie}`
     )
   }
 
@@ -61,8 +62,19 @@ class Welcome extends BasePage {
     )
   }
 
-  async verifyIdentityButton(copy) {
+  async verifyCustomDescriptions() {
+    verifyElementCopy(
+      this.text(),
+      `To open a bank account, we will need to verify your identity.\nIt will only take a couple of minutes.`
+    )
+  }
+
+  async verifyPrimaryButton(copy) {
     verifyElementCopy(this.primaryBtn(), copy.welcome.next_button)
+  }
+
+  async verifyCustomPrimaryButton() {
+    verifyElementCopy(this.primaryBtn(), 'Verify Identity')
   }
 
   async verifyFooter() {