+++ CHANGELOG.md
@@ -14,6 +14,10 @@ This project adheres to the Node [default version scheme](https://docs.npmjs.com
 - Internal: Added note about lockfileVersion and npm version requirement for SDK contributors in CONTRIBUTING.md documentation.
 - Public: Improvements in video capture steps.
 
+### Changed
+
+- UI: Updated default Welcome screen UI
+
 ## Fixed
 
 - Internal: Upgrade Preact from version `10.5.4` to `10.5.13` in order to resolve an unhandled exception on reinitialising the SDK after closing the SDK modal for some integrations when using Modal mode.